ResolveKey: fixed narrowing and null warnings
and made the key param use the same type as its users

+++ gemrb/core/KeyMap.h
@@ -45,7 +45,7 @@ public:
 	KeyMap();
 	~KeyMap();
 	bool InitializeKeyMap(const char *inifile, const char *keyfile);
-	bool ResolveKey(int key, int group);
+	bool ResolveKey(unsigned short key, int group);
 	bool ResolveName(const char* name, int group);
 
 	Function* LookupFunction(const char* name);
